requirements.txt
# Core ML and NLP libraries - Fixed versions for compatibility
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
accelerate==0.25.0

# Document processing
langchain>=0.1.0
langchain-community>=0.0.10
pypdf2>=3.0.1
python-docx>=0.8.11
unstructured>=0.11.0

# Vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Web interface
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
